working towards correcting grid read/write behavior
git-svn-id: https://svn.uni-konstanz.de/disy/jscsi/trunk@465 fd70404c-346f-40f3-aaf5-4d485f690dad

+++ projects/scsi/test/integration/org/jscsi/scsi/transport/TestTargetTransportPort.java
@@ -18,14 +18,14 @@ public class TestTargetTransportPort implements TargetTransportPort
 
    private Random rnd = new Random();
    
-   protected int deviceCapacity;
+   protected long deviceCapacity;
    protected int blockSize;
    
    private HashMap<Long,ByteBuffer> readDataMap = new HashMap<Long,ByteBuffer>();
    private HashMap<Long,ByteBuffer> writeDataMap = new HashMap<Long,ByteBuffer>();
 
    
-   public TestTargetTransportPort(int deviceCapacity, int blockSize)
+   public TestTargetTransportPort(long deviceCapacity, int blockSize)
    {
       this.deviceCapacity = deviceCapacity;
       this.blockSize = blockSize;
